sql-ops: Parallelize the CPU histogram variant for GPU radix partitioning
This change required a rather large refactoring to make the parallel
writing of the prefix sum thread-safe. The way this happens is by
splitting the `RadixPartitionOffsets` into mutable non-overlapping
slices. This ensures that threads write to distinct locations.

In addition, the input is also split into mutable slices.

By mistake I added a mutable slice for `PartitionedRelation`. This was
the most difficult one, and it's not actually used right now. However,
it might be useful in future to integrate the CPU and GPU
implementations, which are currently completely separated.