Microbench: Remove sync benchmark
Broken and unmaintained code. Was originally intended to show that GPU L1 and
CPU L1 caches are coherent. This is not the case, as the GPU L1 is not in any
way coherent.

Note that in future it may become relevant to show the latency of GPU reads
from a CPU cache.  Such a benchmark might reuse some of the now-removed code.