sql-ops: Fix buffer flush for SSWWC and SSWWC_v2 partitioning variants
Buffer may contain invalid data as we align the buffer to cachelines.
The alignment can cause the flush to write the invalid data into the
partition.

Fixed retaining the order of tuples when flushing (instead of randomizing
the order with atomicAdd), and then checking that the destination offset
is actually inside of the partition (instead of before the beginning).