sql-ops: Move offset counters to device memory for SSWWCv2G
Fanout is limited to 2^14 if the offset counters are in shared memory.
Moving them to device memory enables higher fanouts up to 2^20. At this
point, buffers + offsets are about 11 GiB large (the GPU has 16 GiB of
memory).

The performance impact vs. shared memory seems to be negligible. Moving
the buffers already incurred a 10% cost, moving the offsets too doesn't
seem to worsen this impact.