sql-ops: Fix multiply-shift hash function
Multiply-shift hash must be shifted to use the most significant bits in
the output. Cannot just `and` and mask to the required number of buckets.

The result of this change are fewer hash collisions, leading to
increased insert and probe throughput. With both linear probing and
bucket chaining on the CPU and on the GPU, the throughput of
multiply-shift is now faster than Murmur3.