microbench: Measure cold TLB cache in TLB latency bench
The purpose is to trace the exact behavior of the IOTLB. Measurements so
far are very inconsistent.

Due to fine-grained latency measurements, each access is logged
individually. Thus, we can deduce from the data point ID which accesses
are with a cold cache, and which ones are with a warm cache.