numa-gpu: Align pointer to huge pages for mprotect
mprotect fails with error code 22 "invalid argument" if pointer is not
correctly aligned to the page size. This failure occurs for 1 GB huge
pages.

The fix is to pass the page size to the mprotect wrapper. Regular slices
pass in the small page size as before, but huge page allocations pass in
the huge page size.