datagen: Generate primary keys in range [0, max[ instead of [1, max]
Including the maximum value in the key space is an off-by-one error.
This causes a segfault in the perfect hashing, no-partitioned hash join,
when the hash table is exactly aligned on a page boundary.

Test cases that rely on zero not being in the key space should use a
NULL value instead.

The early HSSWWC variants (all except v4) still fail on one test case,
as the device memory buffer is initialized to zero. Won't fix it for
now due to complexity; they should be deleted in future.