radix-join: Enable CPU-partitioned execution method on PCI-e GPUs
Linux memory information is only available for GPU memory if the GPU is
a NUMA node. This is the case for NVLink GPUs, but not for PCI-e GPUs.

Thus, if Linux NumaMemInfo returns an error, then fallback to
CudaMemInfo.

However, a drawback of CudaMemInfo is that it overreports memory usage,
which may lead to out-of-memory failures for large data sets. Don't have
a better solution at the moment.